Breaking: Bump `sharedb` dependency
At the moment, `sharedb-milestone-mongo` has a dependency on a specific
`beta` version of `sharedb`, which means that - when using this library
in an application - the application will probably resolve two different
`sharedb` dependencies.

This change bumps `sharedb-milestone-mongo` to rely on at least `v1.0.0`
of `sharedb`, so that consumer applications can resolve to a unified
`sharedb` version.

Note that the newer version of `sharedb` replaced `expect.js` with
`chai`. Since we run some tests from `sharedb`, this change also updates
our dependencies here to follow suit.